+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+3.12.0 / 2018-08-06
+  * Support hexadecimal values
+  * Support true number parsing 
 3.11.2 / 2018-07-23
   * Update Demo for more options
   * Update license information